+++ src/main/java/oshi/software/os/mac/MacHardwareAbstractionLayer.java
@@ -6,11 +6,13 @@ package oshi.software.os.mac;
 import java.util.ArrayList;
 import java.util.List;
 
+import oshi.hardware.Gpu;
 import oshi.hardware.HardwareAbstractionLayer;
 import oshi.hardware.Memory;
 import oshi.hardware.Processor;
 import oshi.software.os.mac.local.CentralProcessor;
 import oshi.software.os.mac.local.GlobalMemory;
+import oshi.software.os.mac.local.GraphicCard;
 import oshi.util.ExecutingCommand;
 
 /**
@@ -21,6 +23,7 @@ public class MacHardwareAbstractionLayer implements HardwareAbstractionLayer {
 
 	private Processor[] _processors;
 	private Memory _memory;
+	private Gpu[] _gpu;
 
 	/*
 	 * (non-Javadoc)
@@ -41,6 +44,8 @@ public class MacHardwareAbstractionLayer implements HardwareAbstractionLayer {
 
 	}
 
+
+
 	/*
 	 * (non-Javadoc)
 	 * 
@@ -51,6 +56,20 @@ public class MacHardwareAbstractionLayer implements HardwareAbstractionLayer {
 			_memory = new GlobalMemory();
 		}
 		return _memory;
+	}
+
+	public Gpu[] getGpus() {
+		if (_gpu == null) {
+			List<Gpu> gpus = new ArrayList<Gpu>();
+			ArrayList<String> sys_profile = ExecutingCommand.runNative("system_profiler SPDisplaysDataType");
+			for (String line : sys_profile) {
+				if (line.contains("Chipset Model:")) {
+					gpus.add(new GraphicCard(line));
+				}
+			}
+			_gpu = gpus.toArray(new Gpu[gpus.size()]);
+		}
+		return _gpu;
 	}
 
 }