+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Next
 * Added Travis-CI - [@dblock](https://github.com/dblock).
 * Added TODO list and enhanced README documentation - [@ptitvert](https://github.com/ptitvert)
 * [#15](https://github.com/dblock/oshi/pull/15): Added support for CPU load on OSX and Linux - [@kamenitxan](https://github.com/kamenitxan).
+* [#16](https://github.com/dblock/oshi/pull/15): Added support for GPU name and memory on OSX and Linux - [@kamenitxan](https://github.com/kamenitxan).
 * Your contribution here.